+++ partner/src/epilot/partners.py
@@ -259,6 +259,8 @@ class Partners:
     def invite_partner(self, id: str, partner_invitation_payload: Optional[shared.PartnerInvitationPayload] = None) -> operations.InvitePartnerResponse:
         r"""invitePartner
         Create a new partner in partner directory and send an invite email to accept request
+
+        Deprecated method: This will be removed in a future release, please migrate away from it as soon as possible.
         """
         hook_ctx = HookContext(operation_id='invitePartner', oauth2_scopes=[], security_source=self.sdk_configuration.security)
         request = operations.InvitePartnerRequest(
@@ -304,6 +306,71 @@ class Partners:
         
         
         res = operations.InvitePartnerResponse(status_code=http_res.status_code, content_type=http_res.headers.get('Content-Type'), raw_response=http_res)
+        
+        if http_res.status_code == 200:
+            if utils.match_content_type(http_res.headers.get('Content-Type'), 'application/json'):                
+                out = utils.unmarshal_json(http_res.text, Optional[shared.Partner])
+                res.partner = out
+            else:
+                content_type = http_res.headers.get('Content-Type')
+                raise errors.SDKError(f'unknown content-type received: {content_type}', http_res.status_code, http_res.text, http_res)
+        elif http_res.status_code == 400 or http_res.status_code >= 400 and http_res.status_code < 500 or http_res.status_code >= 500 and http_res.status_code < 600:
+            raise errors.SDKError('API error occurred', http_res.status_code, http_res.text, http_res)
+        else:
+            raise errors.SDKError('unknown status code received', http_res.status_code, http_res.text, http_res)
+
+        return res
+
+    
+    
+    def invite_partner_v2(self, id: str, partner_invitation_payload: Optional[shared.PartnerInvitationPayload] = None) -> operations.InvitePartnerV2Response:
+        r"""invitePartnerV2
+        Invite a partner into collaboration. It will send an email to partner and ask to join into collaboration
+        """
+        hook_ctx = HookContext(operation_id='invitePartnerV2', oauth2_scopes=[], security_source=self.sdk_configuration.security)
+        request = operations.InvitePartnerV2Request(
+            id=id,
+            partner_invitation_payload=partner_invitation_payload,
+        )
+        
+        base_url = utils.template_url(*self.sdk_configuration.get_server_details())
+        
+        url = utils.generate_url(operations.InvitePartnerV2Request, base_url, '/v2/partners/{id}/invite', request)
+        
+        if callable(self.sdk_configuration.security):
+            headers, query_params = utils.get_security(self.sdk_configuration.security())
+        else:
+            headers, query_params = utils.get_security(self.sdk_configuration.security)
+        
+        req_content_type, data, form = utils.serialize_request_body(request, operations.InvitePartnerV2Request, "partner_invitation_payload", False, True, 'json')
+        if req_content_type is not None and req_content_type not in ('multipart/form-data', 'multipart/mixed'):
+            headers['content-type'] = req_content_type
+        headers['Accept'] = 'application/json'
+        headers['user-agent'] = self.sdk_configuration.user_agent
+        client = self.sdk_configuration.client
+        
+        try:
+            req = self.sdk_configuration.get_hooks().before_request(
+                hook_ctx, 
+                requests_http.Request('POST', url, params=query_params, data=data, files=form, headers=headers).prepare(),
+            )
+            http_res = client.send(req)
+        except Exception as e:
+            _, e = self.sdk_configuration.get_hooks().after_error(hook_ctx, None, e)
+            raise e
+
+        if utils.match_status_codes(['400','4XX','5XX'], http_res.status_code):
+            http_res, e = self.sdk_configuration.get_hooks().after_error(hook_ctx, http_res, None)
+            if e:
+                raise e
+        else:
+            result = self.sdk_configuration.get_hooks().after_success(hook_ctx, http_res)
+            if isinstance(result, Exception):
+                raise result
+            http_res = result
+        
+        
+        res = operations.InvitePartnerV2Response(status_code=http_res.status_code, content_type=http_res.headers.get('Content-Type'), raw_response=http_res)
         
         if http_res.status_code == 200:
             if utils.match_content_type(http_res.headers.get('Content-Type'), 'application/json'):                
@@ -385,6 +452,8 @@ class Partners:
     def resend_partner_invitation(self, id: str, request_body: Optional[operations.ResendPartnerInvitationRequestBody] = None) -> operations.ResendPartnerInvitationResponse:
         r"""resendPartnerInvitation
         Resend partner invitation email
+
+        Deprecated method: This will be removed in a future release, please migrate away from it as soon as possible.
         """
         hook_ctx = HookContext(operation_id='resendPartnerInvitation', oauth2_scopes=[], security_source=self.sdk_configuration.security)
         request = operations.ResendPartnerInvitationRequest(