+++ partner/src/epilot/sdkconfiguration.py
@@ -23,9 +23,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '4.3.1'
-    gen_version: str = '2.280.6'
-    user_agent: str = 'speakeasy-sdk/python 4.3.1 2.280.6 1.0.0 epilot-partner'
+    sdk_version: str = '4.3.2'
+    gen_version: str = '2.286.7'
+    user_agent: str = 'speakeasy-sdk/python 4.3.2 2.286.7 1.0.0 epilot-partner'
     retry_config: RetryConfig = None
     _hooks: SDKHooks = None